mitosheet: update graph_data schema, reorganize graph api, and other cleanup
