--- mitosheet/mitosheet/api/graph/__init__.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-# Copyright (c) Saga Inc.
-# Distributed under the terms of the GPL License.
-from mitosheet.api.graph.get_column_summary_graph import get_column_summary_graph
