+++ mitosheet/src/api.tsx
@@ -337,8 +337,8 @@ export default class MitoAPI {
         graphType: GraphType,
         sheet_index: number,
         safety_filter_turned_on_by_user: boolean,
-        xAxisColumnIDs: ColumnID[] | undefined,
-        yAxisColumnIDs: ColumnID[] | undefined,
+        xAxisColumnIDs: ColumnID[],
+        yAxisColumnIDs: ColumnID[],
         height: string,
         width: string,
         stepID?: string,
